added logic for vending machine

+++ VendingMachine/Models/Product.cs
@@ -8,10 +8,10 @@ namespace VendingMachine.Models
 {
     public class Product
     {
-        public Dictionary<string, double> ProductsAvailable;
+        public Dictionary<string, decimal> ProductsAvailable;
         public Product()
         {
-            ProductsAvailable = new Dictionary<string, double>() { { "cola" , 1.0 },{ "chips",0.5 },{ "candy",0.65 } };
+            ProductsAvailable = new Dictionary<string, decimal>() { { "cola" , (decimal)1.0 },{ "chips", (decimal)0.5 },{ "candy", (decimal)0.65 } };
 
         }
     }
